src/entities/appointment.spec.ts
import { expect, test } from 'vitest';
import { Appointment } from './appointment';
import { getFutureDate } from '../tests/get-future-date';

test('create an appointment', () => {
    const startsAt = getFutureDate('2024-04-20')
    const endsAt = getFutureDate('2024-04-21')

    const appointment = new Appointment({
        customer: 'John Doe',
        startsAt,
        endsAt
    })

    expect(appointment).toBeInstanceOf(Appointment)
    expect(appointment.customer).toEqual('John Doe')
})

test('cannot create an appointment with end date before start date', () => {
    const startsAt = getFutureDate('2024-04-20')
    const endsAt = getFutureDate('2024-04-19')

    expect(() => {
        return new Appointment({
            customer: 'John Doe',
            startsAt,
            endsAt
        })
    }).toThrow()

})

test('cannot create an appointment with end date before now', () => {
    const startsAt = new Date()
    const endsAt = new Date()

    startsAt.setDate(startsAt.getDate() - 1)
    endsAt.setDate(endsAt.getDate() + 3)

    expect(() => {
        return new Appointment({
            customer: 'John Doe',
            startsAt,
            endsAt
        })
    }).toThrow()

})